pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
#authors = ["ChrisMarSil <chris.mar.silva@gmail.com>"]
authors = ["Admin <admin@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.8.0"}
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.0"
httpx = "^0.27.0"

[tool.ruff]
line-length = 160
indent-width = 4
extend-exclude = ['migrations']
#exclude = [
#    ".bzr",
#    ".direnv",
#    ".eggs",
#    ".git",
#    ".git-rewrite",
#    ".hg",
#    ".ipynb_checkpoints",
#    ".mypy_cache",
#    ".nox",
#    ".pants.d",
#    ".pyenv",
#    ".pytest_cache",
#    ".pytype",
#    ".ruff_cache",
#    ".svn",
#    ".tox",
#    ".venv",
#    ".vscode",
#    "__pypackages__",
#    "_build",
#    "buck-out",
#    "build",
#    "dist",
#    "node_modules",
#    "site-packages",
#    "venv",
#    "migrations",
#]

[tool.ruff.lint]
preview = true
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warning
    "F",  # flake8 error - pyflakes
    "I",  # isort imports
    "B",  # flake8 bugbear
    'PL',
    'PT',
    #"C4",  # flake8-comprehensions # simplify comprehensions
    #"UP",  # pyupgrade # upgade syntax
    #"ARG001", # unused arguments in functions
#    "A",      # builtin shadowing
#    "BLE",    # aboid bare excepts
#    "DTZ",    # datetime errors
#    "FBT",    # avoid boolean trap
#    "G",      # logging format
#    "N",      # conform to PEP8 naming rules
#    "RET",    # return values
#    "S",      # bandit
#    "TRY",    # exceptions antipatterns
#    "YTT",    # wrong usage of sys.info
#    'SIM',
#    "E4",
#    "E7",
#    "E9",
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    #"W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
    # 'F401', 'E402', 'F811'
]

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_mode = "auto"

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev app/api.py'
pre_test = 'task lint'
test = 'pytest --asyncio-mode=auto -s -x --cov=app -vv -x'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
